requirements.txt
django>=5.2,<6.0
pillow>=10.0
django-unfold>=0.18

